+++ CHANGELOG.md
@@ -51,6 +51,7 @@ You can find and compare releases at the [GitHub release page](https://github.co
 - Add `GraphQLContext:: setUser(?Authenticatable $user): void`
 - Add directive `@whereKey` to filter Models by their primary key https://github.com/nuwave/lighthouse/pull/2289
 - Allow directive `@where` on fields https://github.com/nuwave/lighthouse/pull/2306
+- Allow passing route parameters in `MakeGraphQLRequests` test methods https://github.com/nuwave/lighthouse/pull/2333
 
 ### Removed